refactor: Image Parser -> interface, add latest menu test image
