+++ src/test/java/shanepark/foodbox/image/ImageParserTesseractTest.java
@@ -7,7 +7,7 @@ import org.slf4j.LoggerFactory;
 import org.springframework.core.io.ClassPathResource;
 import shanepark.foodbox.image.domain.ParsedMenu;
 import shanepark.foodbox.image.service.ImageMarginCalculator;
-import shanepark.foodbox.image.service.ImageParser;
+import shanepark.foodbox.image.service.ImageParserTesseract;
 
 import java.io.IOException;
 import java.io.InputStream;
@@ -16,14 +16,14 @@ import java.util.List;
 import static org.assertj.core.api.Assertions.assertThat;
 
 
-class ImageParserTest {
-    private final Logger logger = LoggerFactory.getLogger(ImageParserTest.class);
-    ImageParser parser = new ImageParser(new ImageMarginCalculator());
+class ImageParserTesseractTest {
+    private final Logger logger = LoggerFactory.getLogger(ImageParserTesseractTest.class);
+    ImageParserTesseract parser = new ImageParserTesseract(new ImageMarginCalculator());
 
     @Test
     @DisplayName("Parse image and return parsed menu")
     void parse() throws IOException {
-        ClassPathResource resource = new ClassPathResource("menu-nov-11.png");
+        ClassPathResource resource = new ClassPathResource("menu/menu-11112024.png");
         try (InputStream ins = resource.getInputStream()) {
             List<ParsedMenu> parsedMenu = parser.parse(ins);
             logger.info("parsedMenu: \n{}", parsedMenu);
@@ -36,7 +36,7 @@ class ImageParserTest {
     @Test
     @DisplayName("OCT 28 image parse test")
     void parse2() throws IOException {
-        ClassPathResource resource = new ClassPathResource("menu-oct-28.png");
+        ClassPathResource resource = new ClassPathResource("menu/menu-10282024.png");
         try (InputStream ins = resource.getInputStream()) {
             List<ParsedMenu> parsedMenu = parser.parse(ins);
             logger.info("parsedMenu: \n{}", parsedMenu);
@@ -49,7 +49,7 @@ class ImageParserTest {
     @Test
     @DisplayName("DEC 09 image parse test")
     void parse3() throws IOException {
-        ClassPathResource resource = new ClassPathResource("menu-dec-09.png");
+        ClassPathResource resource = new ClassPathResource("menu/menu-12092024-yuseong.png");
         try (InputStream ins = resource.getInputStream()) {
             List<ParsedMenu> parsedMenu = parser.parse(ins);
             logger.info("parsedMenu: \n{}", parsedMenu);
@@ -62,7 +62,7 @@ class ImageParserTest {
     @Test
     @DisplayName("DEC 09 official website image parse test")
     void parse4() throws IOException {
-        ClassPathResource resource = new ClassPathResource("menu-dec-09-official.png");
+        ClassPathResource resource = new ClassPathResource("menu/menu-12092024-official.png");
         try (InputStream ins = resource.getInputStream()) {
             List<ParsedMenu> parsedMenu = parser.parse(ins);
             logger.info("parsedMenu: \n{}", parsedMenu);
