+++ src/test/resources/menu/menu-12092024-official.png
